# specs/openapi.yaml
openapi: 3.0.3
info:
  title: AI 卫星总体设计助手 API
  version: 0.4.0
  description: |
    由 `docs/07-api.md`（v0.2 细化版）生成，用途：**Apifox / Postman 直接导入进行手动测试**。
    - 本地 Base URL：`http://127.0.0.1:8000/api/v1`
    - 认证：`Authorization: Bearer <APP_TOKEN>`（仅 `/healthz` 除外）
    - 任务异步：创建返回 `202`；进度经 SSE（`GET /missions/{id}/events`）或轮询（`GET /missions/{id}`）
    - `STUB_LLM=1` 时为桩模式（离线演示 / CI，全部 LLM 调用走录制回放）
    - 枚举、错误码、SSE 事件全表见 `docs/07-api.md` §1
servers:
  - url: http://127.0.0.1:8000/api/v1
    description: 本地开发
...
security:
  - bearerAuth: []
paths:
...
  /kb/search:
    post:
      operationId: kbSearch
      tags: [KnowledgeBase]
      summary: 检索（B1）
      description: 混合检索（向量 0.7 + 关键词 0.3）→ 元数据过滤 → 返回带引用的结果。
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/KBSearchRequest'
            example:
              query: 小卫星太阳翼面积与功率的典型关系
              filters: { module: power, is_synthetic: true }
              top_k: 8
      responses:
        '200':
          description: 检索结果
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KBSearchResponse'
              example:
                results:
                  - chunk_id: sst-soa-pwr-014-03
                    doc_id: sst-soa-2026-pwr
                    title: SST-SOA 电源章节
                    text: …典型太阳电池阵面积与功率关系…
                    score: 0.83
                    citation:
                      source_url: https://www.nasa.gov/smallsat-institute/sst-soa/
                      page: 14
                      module: power
                      is_synthetic: false
                took_ms: 186
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /kb/ingest:
    post:
      operationId: kbIngest
      tags: [KnowledgeBase]
      summary: 语料摄取（B2）
      description: 存文件 → 解析 → 切分 → embedding → 写 kb_chunks 与 corpus_manifest（同步执行）。
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/KBIngestRequest'
      responses:
        '201':
          description: 摄取完成
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KBIngestResponse'
              example:
                doc_id: sst-soa-2026-pwr
                title: SST-SOA 电源章节
                chunks: 87
                corpus_version: corpus-v2026.10.1
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
  /kb/documents:
    get:
      operationId: listKbDocuments
      tags: [KnowledgeBase]
      summary: 语料列表（B3）
      description: 按 module / doc_type 过滤，按 fetched_at 倒序分页。
      parameters:
        - name: module
          in: query
          schema: { $ref: '#/components/schemas/Module' }
        - name: doc_type
          in: query
          schema: { $ref: '#/components/schemas/DocType' }
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PageSize'
      responses:
        '200':
          description: 列表
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KBDocumentList'
              example:
                items:
                  - doc_id: sst-soa-2026-pwr
                    title: SST-SOA 电源章节
                    module: power
                    doc_type: manual
                    is_synthetic: false
                    license: public-nasa
                    chunks: 87
                    fetched_at: '2026-10-05T08:40:00Z'
                total: 1
                page: 1
                page_size: 20
        '401': { $ref: '#/components/responses/Unauthorized' }
  /kb/documents/{doc_id}:
    parameters:
      - name: doc_id
        in: path
        required: true
        schema: { type: string }
        example: sst-soa-2026-pwr
    delete:
      operationId: deleteKbDocument
      tags: [KnowledgeBase]
      summary: 删除语料（B4）
      description: 必须 confirm=true；级联删除 kb_chunks；写审计；不可恢复。
      parameters:
        - name: confirm
          in: query
          required: true
          schema: { type: boolean, enum: [true] }
      responses:
        '200':
          description: 已删除
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KBDeleteResponse'
              example: { doc_id: sst-soa-2026-pwr, deleted_chunks: 87, corpus_version: corpus-v2026.10.2 }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /kb/stats:
    get:
      operationId: kbStats
      tags: [KnowledgeBase]
      summary: 语料统计（B5）
      description: 统计 corpus_manifest 与 kb_chunks 计数，返回当前语料版本。
      responses:
        '200':
          description: 统计
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KBStats'
              example:
                documents: 42
                chunks: 3120
                corpus_version: corpus-v2026.10.2
                embedding_model: bge-m3
                last_ingest_at: '2026-10-05T08:40:00Z'
        '401': { $ref: '#/components/responses/Unauthorized' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: APP_TOKEN
  parameters:
...
    Page:
      name: page
      in: query
      schema: { type: integer, minimum: 1, default: 1 }
    PageSize:
      name: page_size
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
  responses:
    BadRequest:
      description: 入参不符合 JSON Schema
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error:
              code: SCHEMA_INVALID
              message: constraints.lifetime_years 必须 ≥ 0.5
              details: { field: constraints.lifetime_years }
              trace_id: req-20261005-abc123
    Unauthorized:
      description: Token 缺失或错误
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: UNAUTHORIZED, message: 无效的访问令牌, details: {}, trace_id: req-20261005-abc124 }
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: MISSION_NOT_FOUND, message: 任务不存在, details: { task_id: t-20261005-999 }, trace_id: req-20261005-abc125 }
    Conflict:
      description: 状态冲突（同 409 各错误码：DOC_NOT_READY / STATE_CONFLICT）
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
          example:
            error: { code: STATE_CONFLICT, message: 当前任务状态不允许该操作, details: { status: succeeded }, trace_id: req-20261005-abc126 }
...
  schemas:
...
    Error:
      type: object
      required: [code, message, trace_id]
      properties:
        code:
          type: string
          enum: [SCHEMA_INVALID, UNAUTHORIZED, MISSION_NOT_FOUND, PARAM_NOT_FOUND, DOC_NOT_FOUND,
                 DOC_NOT_READY, STATE_CONFLICT, RATE_LIMITED, PROVIDER_FAILED, INTERNAL]
        message: { type: string }
        details: { type: object, additionalProperties: true }
        trace_id: { type: string }
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error: { $ref: '#/components/schemas/Error' }

    # ---------- 枚举 ----------
...
    DocType:
      type: string
      enum: [standard, manual, report, catalog, lesson, other]
    Module:
      type: string
      enum: [system, power, ttc, data, adcs, orbit, payload, risk, other]

    # ---------- 任务 ----------
...
    Citation:
      type: object
      properties:
        source_url: { type: string, format: uri }
        page: { type: integer, nullable: true }
        module: { allOf: [{ $ref: '#/components/schemas/Module' }], nullable: true }
        is_synthetic: { type: boolean }
    KBSearchFilters:
      type: object
      additionalProperties: false
      properties:
        module: { $ref: '#/components/schemas/Module' }
        doc_type: { $ref: '#/components/schemas/DocType' }
        year_from: { type: integer, minimum: 1950 }
        is_synthetic: { type: boolean, default: true }
    KBSearchRequest:
      type: object
      required: [query]
      additionalProperties: false
      properties:
        query: { type: string, minLength: 1, maxLength: 500 }
        filters: { $ref: '#/components/schemas/KBSearchFilters' }
        top_k: { type: integer, minimum: 1, maximum: 20, default: 8 }
    KBSearchResult:
      type: object
      properties:
        chunk_id: { type: string }
        doc_id: { type: string }
        title: { type: string }
        text: { type: string }
        score: { type: number }
        citation: { $ref: '#/components/schemas/Citation' }
    KBSearchResponse:
      type: object
      properties:
        results: { type: array, items: { $ref: '#/components/schemas/KBSearchResult' } }
        took_ms: { type: integer }
    KBIngestRequest:
      type: object
      required: [file, module, doc_type, license]
      properties:
        file: { type: string, format: binary }
        module: { $ref: '#/components/schemas/Module' }
        doc_type: { $ref: '#/components/schemas/DocType' }
        is_synthetic: { type: boolean, default: false }
        license: { type: string, example: public-nasa }
        title: { type: string }
    KBIngestResponse:
      type: object
      properties:
        doc_id: { type: string }
        title: { type: string }
        chunks: { type: integer }
        corpus_version: { type: string }
    KBDocument:
      type: object
      properties:
        doc_id: { type: string }
        title: { type: string }
        module: { $ref: '#/components/schemas/Module' }
        doc_type: { $ref: '#/components/schemas/DocType' }
        is_synthetic: { type: boolean }
        license: { type: string }
        chunks: { type: integer }
        fetched_at: { type: string, format: date-time }
    KBDocumentList:
      type: object
      required: [items, total, page, page_size]
      properties:
        items: { type: array, items: { $ref: '#/components/schemas/KBDocument' } }
        total: { type: integer }
        page: { type: integer }
        page_size: { type: integer }
    KBDeleteResponse:
      type: object
      properties:
        doc_id: { type: string }
        deleted_chunks: { type: integer }
        corpus_version: { type: string }
    KBStats:
      type: object
      properties:
        documents: { type: integer }
        chunks: { type: integer }
        corpus_version: { type: string }
        embedding_model: { type: string }
        last_ingest_at: { type: string, format: date-time }

    # ---------- 系统 ----------
...